+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Set-SsisParameter command.
+- Get-SsisParameter command.
+- Remove-SsisEnvironmentReference command.
+- New-SsisEnvironmentReference command.
+- Get-SsisEnvironmentReference command.
 - Remove-SsisEnvironmentVariable command.
 - Set-SsisEnvironmentVariable command.
 - Get-SsisEnvironmentVariable command.